AV-85019 #comment Change the GDP CRD to the new format, update the clientsets

+++ gslb/ingestion/gdp_controller.go
@@ -17,23 +17,29 @@ package ingestion
 import (
 	filter "amko/gslb/gdp_filter"
 	"amko/gslb/gslbutils"
+	"amko/gslb/k8sobjects"
 	"errors"
 	"strconv"
 
-	gdpalphav1 "amko/pkg/apis/avilb/v1alpha1"
+	gdpalphav1 "amko/pkg/apis/amko/v1alpha1"
 	gslbcs "amko/pkg/client/clientset/versioned"
 	gdpscheme "amko/pkg/client/clientset/versioned/scheme"
 	gslbinformers "amko/pkg/client/informers/externalversions"
-	gdplisters "amko/pkg/client/listers/avilb/v1alpha1"
+	gdplisters "amko/pkg/client/listers/amko/v1alpha1"
 
 	"github.com/avinetworks/container-lib/utils"
 	"github.com/openshift/client-go/route/clientset/versioned/scheme"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/kubernetes"
 	typedcorev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/client-go/util/workqueue"
+)
+
+const (
+	AlreadyExistsErr = "a GDP object already exists, can't add another"
 )
 
 // GDPAddDelfn is a type of function which handles an add or a delete of a GDP
@@ -60,6 +66,32 @@ func (gdpController *GDPController) Run(stopCh <-chan struct{}) error {
 	<-stopCh
 	gslbutils.Logf("object: GDPController, msg: %s", "shutting down the workers")
 	return nil
+}
+
+func AddOrUpdateNSStore(clusterNSStore *gslbutils.ObjectStore, ns *corev1.Namespace, cname string) {
+	nsMeta := k8sobjects.GetNSMeta(ns, cname)
+	clusterNSStore.AddOrUpdate(cname, nsMeta.Name, nsMeta)
+}
+
+func DeleteFromNSStore(clusterNSStore *gslbutils.ObjectStore, ns *corev1.Namespace, cname string) {
+	clusterNSStore.DeleteNSObj(cname, ns.Name)
+}
+
+func MoveNSObjs(nsObjs []string, fromStore *gslbutils.ObjectStore, toStore *gslbutils.ObjectStore) {
+	var cname, ns string
+	var err error
+	for _, multiClusterNS := range nsObjs {
+		cname, ns, err = gslbutils.SplitMultiClusterNS(multiClusterNS)
+		if err != nil {
+			gslbutils.Errf("objType: Namespace, object: %s, msg: processing error %s", ns, err.Error())
+			continue
+		}
+		obj, ok := fromStore.DeleteNSObj(cname, ns)
+		if ok {
+			// Object was found, add this to "toStore"
+			toStore.AddOrUpdate(cname, ns, obj)
+		}
+	}
 }
 
 // MoveObjs moves the objects in "objList" from "fromStore" to "toStore".
@@ -108,12 +140,10 @@ func splitName(objType, objName string) (string, string, string, error) {
 	return cname, ns, sname, err
 }
 
-func writeChangedObjToQueue(objType string, k8swq []workqueue.RateLimitingInterface, numWorkers uint32, trafficWeightChanged bool) {
+func GetObjTypeStores(objType string) (string, *gslbutils.ClusterStore, *gslbutils.ClusterStore, error) {
+	var objKey string
 	var acceptedObjStore *gslbutils.ClusterStore
 	var rejectedObjStore *gslbutils.ClusterStore
-	var objKey string
-	var cname, ns, sname string
-	var err error
 
 	if objType == gdpalphav1.RouteObj {
 		acceptedObjStore = gslbutils.GetAcceptedRouteStore()
@@ -129,14 +159,25 @@ func writeChangedObjToQueue(objType string, k8swq []workqueue.RateLimitingInterf
 		objKey = gslbutils.IngressType
 	} else {
 		gslbutils.Warnf("Unknown Object type: %s", objType)
+		return "", nil, nil, errors.New("unknown object type " + objType)
+	}
+	return objKey, acceptedObjStore, rejectedObjStore, nil
+}
+
+func writeChangedObjToQueue(objType string, k8swq []workqueue.RateLimitingInterface, numWorkers uint32, trafficWeightChanged bool) {
+	var cname, ns, sname string
+	var err error
+
+	objKey, acceptedObjStore, rejectedObjStore, err := GetObjTypeStores(objType)
+	if err != nil {
+		gslbutils.Errf("objtype error: %s", err.Error())
 		return
 	}
-	gf := filter.GetGlobalFilter()
 	if acceptedObjStore != nil {
 		// If we have objects in the accepted store, each one has to be passed through
 		// the filter again. If any object fails to pass through the filter, we need to
 		// add DELETE keys for them.
-		acceptedList, rejectedList := acceptedObjStore.GetAllFilteredClusterNSObjects(gf.ApplyFilter)
+		acceptedList, rejectedList := acceptedObjStore.GetAllFilteredClusterNSObjects(filter.ApplyFilter)
 		if len(rejectedList) != 0 {
 			gslbutils.Logf("ObjList: %v, msg: %s", rejectedList, "obj list will be deleted")
 			// Since, these objects are now rejected, they have to be moved to
@@ -158,17 +199,19 @@ func writeChangedObjToQueue(objType string, k8swq []workqueue.RateLimitingInterf
 			}
 		}
 		// if the traffic weight changed, then the accepted list has to be sent to the nodes layer
-		for _, objName := range acceptedList {
-			cname, ns, sname, err = splitName(objType, objName)
-			if err != nil {
-				gslbutils.Errf("msg: couldn't split the key: %s, error, %s", objName, err)
-				continue
+		if trafficWeightChanged {
+			for _, objName := range acceptedList {
+				cname, ns, sname, err = splitName(objType, objName)
+				if err != nil {
+					gslbutils.Errf("msg: couldn't split the key: %s, error, %s", objName, err)
+					continue
+				}
+				bkt := utils.Bkt(ns, numWorkers)
+				key := gslbutils.MultiClusterKey(gslbutils.ObjectUpdate, objKey, cname, ns, sname)
+				k8swq[bkt].AddRateLimited(key)
+				gslbutils.Logf("cluster: %s, ns: %s, objtype: %s, name: %s, key: %s, msg: added key",
+					cname, ns, objType, sname, key)
 			}
-			bkt := utils.Bkt(ns, numWorkers)
-			key := gslbutils.MultiClusterKey(gslbutils.ObjectUpdate, objKey, cname, ns, sname)
-			k8swq[bkt].AddRateLimited(key)
-			gslbutils.Logf("cluster: %s, ns: %s, objtype: %s, name: %s, key: %s, msg: added key",
-				cname, ns, objType, sname, key)
 		}
 	}
 
@@ -176,7 +219,7 @@ func writeChangedObjToQueue(objType string, k8swq []workqueue.RateLimitingInterf
 		// If we have objects in the rejected store, each one has to be passed through
 		// the filter again. If any object passes through the filter, we need to add ADD
 		// keys for them.
-		acceptedList, _ := rejectedObjStore.GetAllFilteredClusterNSObjects(gf.ApplyFilter)
+		acceptedList, _ := rejectedObjStore.GetAllFilteredClusterNSObjects(filter.ApplyFilter)
 		if len(acceptedList) != 0 {
 			gslbutils.Logf("ObjList: %v, msg: %s", acceptedList, "object list will be added")
 			MoveObjs(acceptedList, rejectedObjStore, acceptedObjStore, objKey)
@@ -203,44 +246,37 @@ func validObjectType(objType string) bool {
 	return false
 }
 
-func validOperation(op string) bool {
-	if op == gdpalphav1.EqualsOp || op == gdpalphav1.NotequalsOp {
-		return true
-	}
-	return false
-}
-
-func validLabel(label gdpalphav1.Label) error {
-	if label.Key == "" {
-		if label.Value == "" {
-			return errors.New("label key and value is missing")
+func validLabel(label map[string]string) error {
+	for k, v := range label {
+		if k == "" {
+			return errors.New("label value is missing for key " + k)
 		}
-		return errors.New("label key is missing for value " + label.Value)
-	}
-	if label.Value == "" {
-		return errors.New("label value is missing for key " + label.Key)
+		if v == "" {
+			return errors.New("label key is missing for value " + v)
+		}
 	}
 	return nil
 }
 
 func GDPSanityChecks(gdp *gdpalphav1.GlobalDeploymentPolicy) error {
-	// MatchRule checks
-	for _, matchRule := range gdp.Spec.MatchRules {
-		if !validObjectType(matchRule.Object) {
-			return errors.New("unsupported object type " + matchRule.Object)
+	// MatchRules checks
+	mr := gdp.Spec.MatchRules
+	// no app selector and no namespace selector means, no objects selected
+	if len(mr.AppSelector.Label) > 0 {
+		if err := validLabel(mr.AppSelector.Label); err != nil {
+			return errors.New(err.Error() + " for appSelector")
 		}
-		if !validOperation(matchRule.Op) {
-			return errors.New("unsupported operation " + matchRule.Op)
-		}
-		if err := validLabel(matchRule.Label); err != nil {
-			return errors.New(err.Error())
+	}
+	if len(mr.NamespaceSelector.Label) > 0 {
+		if err := validLabel(mr.NamespaceSelector.Label); err != nil {
+			return errors.New(err.Error() + "for namespaceSelector")
 		}
 	}
 
 	// MatchClusters checks, empty matchClusters are allowed
-	for _, matchCluster := range gdp.Spec.MatchClusters {
-		if !gslbutils.IsClusterContextPresent(matchCluster.ClusterContext) {
-			return errors.New("cluster context " + matchCluster.ClusterContext + " not present in GSLBConfig")
+	for _, cluster := range gdp.Spec.MatchClusters {
+		if !gslbutils.IsClusterContextPresent(cluster) {
+			return errors.New("cluster context " + cluster + " not present in GSLBConfig")
 		}
 	}
 
@@ -264,19 +300,111 @@ func updateGDPStatus(gdp *gdpalphav1.GlobalDeploymentPolicy, msg string) {
 	if !gslbutils.PublishGDPStatus {
 		return
 	}
-	obj, updateErr := gslbutils.GlobalGslbClient.AvilbV1alpha1().GlobalDeploymentPolicies(gdp.GetObjectMeta().GetNamespace()).Update(gdp)
+	obj, updateErr := gslbutils.GlobalGslbClient.AmkoV1alpha1().GlobalDeploymentPolicies(gdp.GetObjectMeta().GetNamespace()).Update(gdp)
 	if updateErr != nil {
 		gslbutils.Errf("Error in updating the GDP status object %v: %s", obj, updateErr)
 	}
 }
 
-// AddGDPObj creates a new GlobalFilter if not present on the first GDP object. Subsequent ADD calls add
-// on to the existing GlobalFilter. For each namespace, there can only be one filter. So, if a filter
-// already exists for a namespace, a user needs to edit that and not add a new one. This rule is taken
-// care of in the admission controller. All in all, a namespace can have only one GDP object and hence,
-// only one filter object.
+func filterExists(f *gslbutils.GlobalFilter) bool {
+	// Check if a filter already exists for this namespace
+	// Check if AppFilter or NSFilter is set
+	if f != nil {
+		f.GlobalLock.RLock()
+		defer f.GlobalLock.RUnlock()
+		if f.AppFilter != nil {
+			return true
+		}
+		if f.NSFilter != nil {
+			f.NSFilter.Lock.RLock()
+			defer f.NSFilter.Lock.RUnlock()
+			if len(f.NSFilter.SelectedNS) > 0 {
+				return true
+			}
+		}
+	}
+	// For all other conditions, return false
+	return false
+}
+
+func WriteChangedObjsToQueue(k8swq []workqueue.RateLimitingInterface, numWorkers uint32, trafficWeightChanged bool) {
+	writeChangedObjToQueue(gdpalphav1.RouteObj, k8swq, numWorkers, trafficWeightChanged)
+	writeChangedObjToQueue(gdpalphav1.LBSvcObj, k8swq, numWorkers, trafficWeightChanged)
+	writeChangedObjToQueue(gdpalphav1.IngressObj, k8swq, numWorkers, trafficWeightChanged)
+}
+
+func applyAndUpdateNamespaces() {
+	acceptedNSStore := gslbutils.GetAcceptedNSStore()
+	rejectedNSStore := gslbutils.GetRejectedNSStore()
+
+	// first move from acceptedStore to rejectedStore
+	gslbutils.Logf("applying filter on all rejected namespaces")
+	_, rejectedList := acceptedNSStore.GetAllFilteredNamespaces(filter.ApplyFilter)
+	if len(rejectedList) != 0 {
+		gslbutils.Logf("objList: %v, msg: obj list will be deleted", rejectedList)
+		MoveNSObjs(rejectedList, acceptedNSStore, rejectedNSStore)
+		// we also need to delete these namespaces from the filter
+		for _, objName := range rejectedList {
+			cname, ns, err := gslbutils.SplitMultiClusterNS(objName)
+			if err != nil {
+				gslbutils.Errf("cluster: %s, ns: %s, msg: key processing error", cname, ns)
+				continue
+			}
+			nsMetaObj, ok := rejectedNSStore.GetNSObjectByName(cname, ns)
+			if !ok {
+				// object doesn't exist, continue
+				gslbutils.Warnf("objName: namespace, msg: object doesn't exist in the rejected store, returning")
+				continue
+			}
+			nsMeta := nsMetaObj.(k8sobjects.NSMeta)
+			nsMeta.DeleteFromFilter()
+		}
+	}
+
+	acceptedList, _ := rejectedNSStore.GetAllFilteredNamespaces(filter.ApplyFilter)
+	if len(acceptedList) != 0 {
+		gslbutils.Logf("objList: %v, msg: obj list will be added", acceptedList)
+		MoveNSObjs(acceptedList, rejectedNSStore, acceptedNSStore)
+		// no need to add these namespaces to the filter, as they are already added via the ApplyFilter() function
+	}
+}
+
+func applyAndAcceptNamespaces() {
+	acceptedNSStore := gslbutils.GetAcceptedNSStore()
+	rejectedNSStore := gslbutils.GetRejectedNSStore()
+
+	// Since, we have just added a fresh GDP object, all the previous namespaces will be in rejected store
+	// so, apply and move the objects from rejected store
+	acceptedList, rejectedList := rejectedNSStore.GetAllFilteredNamespaces(filter.ApplyFilter)
+	if len(rejectedList) == 0 {
+		gslbutils.Logf("rejected list of namespaces is empty, nothing to be done")
+		return
+	}
+
+	MoveNSObjs(acceptedList, rejectedNSStore, acceptedNSStore)
+	gslbutils.Logf("objList: %v, msg: moved these namespaces from rejected to accepted store", acceptedList)
+}
+
+// AddGDPObj creates a new GlobalFilter if not present on the first GDP object. Subsequent
+// adds for GDP objects must fail as only one GDP object is allowed globally.
 func AddGDPObj(obj interface{}, k8swq []workqueue.RateLimitingInterface, numWorkers uint32) {
-	gdp := obj.(*gdpalphav1.GlobalDeploymentPolicy)
+	gdp, ok := obj.(*gdpalphav1.GlobalDeploymentPolicy)
+	if !ok {
+		gslbutils.Errf("object added is not of type GDP")
+		return
+	}
+
+	// GDPs for all other namespaces are rejected
+	if gdp.ObjectMeta.Namespace != gslbutils.AVISystem {
+		return
+	}
+
+	// If amko is restarted, we have to ignore all the other GDP objects which were previously rejected
+	if gdp.Status.ErrorStatus == AlreadyExistsErr {
+		gslbutils.Errf("A GDP object already exists, can't process this object")
+		return
+	}
+
 	err := GDPSanityChecks(gdp)
 	if err != nil {
 		gslbutils.Errf("Error in accepting GDP object: %s", err.Error())
@@ -284,11 +412,11 @@ func AddGDPObj(obj interface{}, k8swq []workqueue.RateLimitingInterface, numWork
 		return
 	}
 
-	// Check if a filter already exists for this namespace
-	gf := filter.GetGlobalFilter()
-	if gf.IsNSFilterExists(gdp.ObjectMeta.GetNamespace()) {
-		gslbutils.Errf("a GDP object already exists for this namespace %s, can't add another", gdp.ObjectMeta.Namespace)
-		updateGDPStatus(gdp, "a GDP object already exists for namespace "+gdp.ObjectMeta.GetNamespace())
+	gf := gslbutils.GetGlobalFilter()
+	if filterExists(gf) {
+		msg := "a GDP object already exists, can't add another"
+		gslbutils.Errf(msg)
+		updateGDPStatus(gdp, msg)
 		return
 	}
 	updateGDPStatus(gdp, "success")
@@ -297,10 +425,10 @@ func AddGDPObj(obj interface{}, k8swq []workqueue.RateLimitingInterface, numWork
 		"GDP object added")
 
 	gslbutils.Logf("creating a new filter")
-	gf.AddToGlobalFilter(gdp)
-	writeChangedObjToQueue(gdpalphav1.RouteObj, k8swq, numWorkers, false)
-	writeChangedObjToQueue(gdpalphav1.LBSvcObj, k8swq, numWorkers, false)
-	writeChangedObjToQueue(gdpalphav1.IngressObj, k8swq, numWorkers, false)
+	gf.AddToFilter(gdp)
+	// First apply the filter on the namespaces
+	applyAndAcceptNamespaces()
+	WriteChangedObjsToQueue(k8swq, numWorkers, false)
 }
 
 // UpdateGDPObj updates the global and the namespace filters if a the GDP object
@@ -315,6 +443,11 @@ func UpdateGDPObj(old, new interface{}, k8swq []workqueue.RateLimitingInterface,
 		return
 	}
 
+	if oldGdp.Status.ErrorStatus == AlreadyExistsErr {
+		gslbutils.Errf("A GDP object already exists, updates will be ignored for other GDP objects")
+		return
+	}
+
 	err := GDPSanityChecks(newGdp)
 	if err != nil {
 		gslbutils.Errf("Error in accepting the new GDP object: %s", err.Error())
@@ -323,7 +456,7 @@ func UpdateGDPObj(old, new interface{}, k8swq []workqueue.RateLimitingInterface,
 	}
 	updateGDPStatus(newGdp, "success")
 
-	gf := filter.GetGlobalFilter()
+	gf := gslbutils.GetGlobalFilter()
 	if gf == nil {
 		// global filter not initialized, return
 		gslbutils.Errf("object: GlobalFilter, msg: global filter not initialized, can't update")
@@ -331,9 +464,9 @@ func UpdateGDPObj(old, new interface{}, k8swq []workqueue.RateLimitingInterface,
 	}
 	if gdpChanged, trafficWeightChanged := gf.UpdateGlobalFilter(oldGdp, newGdp); gdpChanged {
 		gslbutils.Logf("GDP object changed, will go through the objects again")
-		writeChangedObjToQueue(gdpalphav1.RouteObj, k8swq, numWorkers, trafficWeightChanged)
-		writeChangedObjToQueue(gdpalphav1.LBSvcObj, k8swq, numWorkers, trafficWeightChanged)
-		writeChangedObjToQueue(gdpalphav1.IngressObj, k8swq, numWorkers, trafficWeightChanged)
+		// first apply and update the namespaces in the filter
+		applyAndUpdateNamespaces()
+		WriteChangedObjsToQueue(k8swq, numWorkers, trafficWeightChanged)
 	}
 }
 
@@ -345,16 +478,20 @@ func DeleteGDPObj(obj interface{}, k8swq []workqueue.RateLimitingInterface, numW
 	gdp := obj.(*gdpalphav1.GlobalDeploymentPolicy)
 	gslbutils.Logf("ns: %s, gdp: %s, msg: %s", gdp.ObjectMeta.Namespace, gdp.ObjectMeta.Name,
 		"deleted GDP object")
-	gf := filter.GetGlobalFilter()
+	if gdp.Status.ErrorStatus == AlreadyExistsErr {
+		gslbutils.Errf("A GDP object already exists, deletes to other objects won't have any effect")
+		return
+	}
+
+	gf := gslbutils.GetGlobalFilter()
 	if gf == nil {
 		gslbutils.Errf("object: GlobalFilter, msg: global filter not initialized, can't delete")
 		return
 	}
 	gf.DeleteFromGlobalFilter(gdp)
-	// Need to re-evaluate the objects again according to the deleted filter
-	writeChangedObjToQueue(gdpalphav1.RouteObj, k8swq, numWorkers, false)
-	writeChangedObjToQueue(gdpalphav1.LBSvcObj, k8swq, numWorkers, false)
-	writeChangedObjToQueue(gdpalphav1.IngressObj, k8swq, numWorkers, false)
+	// remove all namespaces from filter and re-apply
+	k8sobjects.RemoveAllSelectedNamespaces()
+	WriteChangedObjsToQueue(k8swq, numWorkers, false)
 }
 
 // InitializeGDPController handles initialization of a controller which handles
@@ -365,7 +502,7 @@ func InitializeGDPController(kubeclientset kubernetes.Interface,
 	AddGDPFunc GDPAddDelfn, UpdateGDPFunc GDPUpdfn,
 	DeleteGDPFunc GDPAddDelfn) *GDPController {
 
-	gdpInformer := gslbInformerFactory.Avilb().V1alpha1().GlobalDeploymentPolicies()
+	gdpInformer := gslbInformerFactory.Amko().V1alpha1().GlobalDeploymentPolicies()
 	gdpscheme.AddToScheme(scheme.Scheme)
 	gslbutils.Logf("object: GDPController, msg: %s", "creating event broadcaster")
 	eventBroadcaster := record.NewBroadcaster()
